# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Primary School Website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NewsArticle {
  id: number;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  imageUrl: string;
  publishedAt: string;
}

export interface NewsArticleInput {
  title: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  imageUrl: string;
}

export interface SchoolEvent {
  id: number;
  title: string;
  description: string;
  date: string;
  time: string;
  location: string;
  category: string;
}

export interface StaffMember {
  id: number;
  name: string;
  role: string;
  department: string;
  bio: string;
  photoUrl: string;
  displayOrder: number;
}

export interface GalleryItem {
  id: number;
  title: string;
  imageUrl: string;
  category: string;
  createdAt: string;
}

export type ListNewsParams = {
limit?: number;
category?: string;
};

export type ListEventsParams = {
limit?: number;
};

export type ListStaffParams = {
department?: string;
};

export type ListGalleryParams = {
category?: string;
limit?: number;
};