# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Primary School Website API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: news
    description: News articles
  - name: events
    description: School events
  - name: staff
    description: Staff members
  - name: gallery
    description: Photo gallery
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of news articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
    post:
      operationId: createNews
      tags: [news]
      summary: Create a news article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"

  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get a news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List upcoming events
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SchoolEvent"

  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      parameters:
        - name: department
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"

  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    NewsArticle:
      type: object
      required: [id, title, excerpt, content, category, author, imageUrl, publishedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        imageUrl:
          type: string
        publishedAt:
          type: string

    NewsArticleInput:
      type: object
      required: [title, excerpt, content, category, author, imageUrl]
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        imageUrl:
          type: string

    SchoolEvent:
      type: object
      required: [id, title, description, date, time, location, category]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: string
        category:
          type: string

    StaffMember:
      type: object
      required: [id, name, role, department, bio, photoUrl, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        department:
          type: string
        bio:
          type: string
        photoUrl:
          type: string
        displayOrder:
          type: integer

    GalleryItem:
      type: object
      required: [id, title, imageUrl, category, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        createdAt:
          type: string