# lib/api-zod/src/generated/types/galleryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Primary School Website API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryItem {
  id: number;
  title: string;
  imageUrl: string;
  category: string;
  createdAt: string;
}