# lib/api-zod/src/generated/types/newsArticleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Primary School Website API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsArticleInput {
  title: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  imageUrl: string;
}