# lib/api-zod/src/generated/types/schoolEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Primary School Website API
 * OpenAPI spec version: 0.1.0
 */

export interface SchoolEvent {
  id: number;
  title: string;
  description: string;
  date: string;
  time: string;
  location: string;
  category: string;
}