# lib/api-zod/src/generated/types/staffMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Primary School Website API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffMember {
  id: number;
  name: string;
  role: string;
  department: string;
  bio: string;
  photoUrl: string;
  displayOrder: number;
}